backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
joblib
numpy
pandas
requests
scikit-learn
gunicorn
pytest
